Move script helper with other asset helpers.
